requirements.txt
fastapi
uvicorn
python-dotenv
openai
line-bot-sdk
google-cloud-firestore